fix: Update types

+++ source-data/era_5.py
@@ -16,7 +16,7 @@ client = Client(url=url, key=key)
 class DownloadData(models.DataDownloadBase):
     def __init__(
         self,
-        location_coord: tuple[int],
+        location_coord: tuple[float],
         aggregation: models.AggregationLevel,
         date_from_utc: date,
         date_to_utc: date,
